pyproject.toml
[project]
name = "access-ci-account-api"
version = "0.2.0"
description = "API for ACCESS CI accounts and registration"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "boto3>=1.41.2",
    "fastapi[standard]>=0.120.2",
    "httpx>=0.28.1",
    "PyJWT>=2.8.0",
    "python-multipart>=0.0.9",
]
